Creating a ground component
