+++ client/src/App.tsx
@@ -9,7 +9,7 @@ import Overlay from "./components/Overlay";
 import CarryTracker from "./components/CarryTracker";
 import SettingsControls from "./components/SettingsControls";
 import { OrbitControls, Sky } from "@react-three/drei";
-import { TextureLoader } from "three";
+import Ground from "./components/Ground";
 
 export default function App() {
   const theme = useTheme();
@@ -51,29 +51,6 @@ export default function App() {
         </Overlay>
       </Box>
     </MantineProvider>
-  );
-}
-
-function Ground() {
-  const texture = useLoader(TextureLoader, "textures/grasslight-big.jpeg");
-
-  if (texture) {
-    texture.wrapS = texture.wrapT = THREE.RepeatWrapping;
-    texture.repeat.set(1500, 1500);
-    texture.anisotropy = 16;
-  }
-
-  return (
-    <mesh receiveShadow position={[0, -1, 0]} rotation={[Math.PI / 2, 0, 0]}>
-      <planeBufferGeometry attach="geometry" args={[10000, 10000]} />
-      {texture && (
-        <meshPhongMaterial
-          attach="material"
-          map={texture}
-          side={THREE.DoubleSide}
-        />
-      )}
-    </mesh>
   );
 }
